Java fully cooked
Java code debugged and cleaned up, verfied to run. JVM -Xss variable set
to 1000000k to ensure that massive recursion doesn't cause a stack
exception on the Recursive Test (for very large values of DEPTH). This
is similar to the Stack/Commit Memory Size parameters set on the C++ and
C++/CLI projects (it seems C# can handle that itself dynamically).